added export defaults, and began battle page build
