+++ public/battle.html
@@ -1,0 +1,104 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="./assets/css/battle.css">
+    <script defer src="./assets/js/battle.js"></script>
+    <title>Dungeons & Devs</title>
+</head>
+<body>
+    <header id="title-container">
+        <div id="logo-container">
+
+        </div>
+        <div id="username-contianer">
+            <h2 id="playerUsername">PlayerUsername</h2>
+        </div>
+    </header>
+
+    <main id="gameboard-container">
+        <section id="playerCard-container">
+            <div id="playerName-container">
+                <h3 id="playerName">PlayerName</h3>
+            </div>
+            <div id="attack-container">
+                <p id="attackVal">7</p>
+            </div>
+            <div id="defense-container">
+                <p id="defenseVal">6</p>
+            </div>
+            <div id="evasion-container">
+                <p id="evasionVal">7</p>
+            </div>
+        </section>
+        <div id="playerHealthbar-container">
+            <div id="healthLogo"></div>
+            <div id="playerHealthbar">
+                <div id="playerHealthCurrent">
+
+                </div>
+            </div>
+        </div>
+
+        <section id="bossCard-container">
+            <div id="bossName-container">
+                <h3 id="playerName">BossName</h3>
+            </div>
+            <div id="attack-container">
+                <p id="attackVal">8</p>
+            </div>
+            <div id="defense-container">
+                <p id="defenseVal">6</p>
+            </div>
+            <div id="evasion-container">
+                <p id="evasionVal">6</p>
+            </div>
+        </section>
+        <div id="BossHealthbar-container">
+            <div id="healthLogo"></div>
+            <div id="BossHealthbar">
+                <div id="BossHealthCurrent">
+
+                </div>
+            </div>
+        </div>
+
+        <aside id="Card1">
+            <div id="card1Title-container">
+                <h4 id="card1Title">Attack</h4>
+            </div>
+            <div id="card1Icon">
+
+            </div>
+            <div id="card1Val-container">
+                <h5 id="card1Val">+5</h5>
+            </div>
+        </aside>
+        <aside id="Card2">
+            <div id="card2Title-container">
+                <h4 id="card2Title">Evasion</h4>
+            </div>
+            <div id="card2Icon">
+
+            </div>
+            <div id="card2Val-container">
+                <h5 id="card2Val">+2</h5>
+            </div>
+        </aside>
+        <aside id="Card3">
+            <div id="card3Title-container">
+                <h4 id="card3Title">Defense</h4>
+            </div>
+            <div id="card3Icon">
+
+            </div>
+            <div id="card3Val-container">
+                <h5 id="card3Val">+3</h5>
+            </div>
+        </aside>
+    </main>
+    
+</body>
+</html>
